// src/typechain/ITierV2.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import {
  BaseContract,
  BigNumber,
  BigNumberish,
  BytesLike,
  CallOverrides,
  PopulatedTransaction,
  Signer,
  utils,
} from "ethers";
import { FunctionFragment, Result } from "@ethersproject/abi";
import { Listener, Provider } from "@ethersproject/providers";
import { TypedEventFilter, TypedEvent, TypedListener, OnEvent } from "./common";

export interface ITierV2Interface extends utils.Interface {
  functions: {
    "report(address,uint256[])": FunctionFragment;
    "reportTimeForTier(address,uint256,uint256[])": FunctionFragment;
  };

  encodeFunctionData(
    functionFragment: "report",
    values: [string, BigNumberish[]]
  ): string;
  encodeFunctionData(
    functionFragment: "reportTimeForTier",
    values: [string, BigNumberish, BigNumberish[]]
  ): string;

  decodeFunctionResult(functionFragment: "report", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "reportTimeForTier",
    data: BytesLike
  ): Result;

  events: {};
}

export interface ITierV2 extends BaseContract {
  connect(signerOrProvider: Signer | Provider | string): this;
  attach(addressOrName: string): this;
  deployed(): Promise<this>;

  interface: ITierV2Interface;

  queryFilter<TEvent extends TypedEvent>(
    event: TypedEventFilter<TEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TEvent>>;

  listeners<TEvent extends TypedEvent>(
    eventFilter?: TypedEventFilter<TEvent>
  ): Array<TypedListener<TEvent>>;
  listeners(eventName?: string): Array<Listener>;
  removeAllListeners<TEvent extends TypedEvent>(
    eventFilter: TypedEventFilter<TEvent>
  ): this;
  removeAllListeners(eventName?: string): this;
  off: OnEvent<this>;
  on: OnEvent<this>;
  once: OnEvent<this>;
  removeListener: OnEvent<this>;

  functions: {
    report(
      account: string,
      context: BigNumberish[],
      overrides?: CallOverrides
    ): Promise<[BigNumber] & { report: BigNumber }>;

    reportTimeForTier(
      account: string,
      tier: BigNumberish,
      context: BigNumberish[],
      overrides?: CallOverrides
    ): Promise<[BigNumber] & { time: BigNumber }>;
  };

  report(
    account: string,
    context: BigNumberish[],
    overrides?: CallOverrides
  ): Promise<BigNumber>;

  reportTimeForTier(
    account: string,
    tier: BigNumberish,
    context: BigNumberish[],
    overrides?: CallOverrides
  ): Promise<BigNumber>;

  callStatic: {
    report(
      account: string,
      context: BigNumberish[],
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    reportTimeForTier(
      account: string,
      tier: BigNumberish,
      context: BigNumberish[],
      overrides?: CallOverrides
    ): Promise<BigNumber>;
  };

  filters: {};

  estimateGas: {
    report(
      account: string,
      context: BigNumberish[],
      overrides?: CallOverrides
    ): Promise<BigNumber>;

    reportTimeForTier(
      account: string,
      tier: BigNumberish,
      context: BigNumberish[],
      overrides?: CallOverrides
    ): Promise<BigNumber>;
  };

  populateTransaction: {
    report(
      account: string,
      context: BigNumberish[],
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;

    reportTimeForTier(
      account: string,
      tier: BigNumberish,
      context: BigNumberish[],
      overrides?: CallOverrides
    ): Promise<PopulatedTransaction>;
  };
}